fix(wextra): fix pr-gate failures (clang-tidy + asan build)
clang-tidy: add NOLINT(readability-redundant-member-init) to = {}
initializers added for -Wmissing-designated-field-initializers; these
are intentional and must not be removed.

asan build / sign-compare (-Werror with clang):
- prepare_conv2d_weights.cpp: change int/auto loop vars to uint32_t
  where compared against uint32_t shape array elements (lines 242-243,
  592-593, 660-672)
- test_unit_mesh_utils.cpp: int -> size_t loop var vs .size()

asan build / -Wmissing-designated-field-initializers:
- mesh_device_operation_adapter.hpp:500: add explicit
  .workload_descriptor = {} to designated initializer that was only
  setting .resolved_bindings

+++ tests/ttnn/unit_tests/gtests/tensor/test_unit_mesh_utils.cpp
@@ -131,7 +131,7 @@ TEST_F(UnitMeshUtils2x4Test, AggregateMismatchedTensorSpecs) {
 
     std::vector<Tensor> tensors;
     tensors.reserve(unit_meshes.size());
-    for (int i = 0; i < unit_meshes.size(); i++) {
+    for (size_t i = 0; i < unit_meshes.size(); i++) {
         if (i % 2 == 0) {
             tensors.push_back(create_device_tensor(
                 tt::tt_metal::TensorSpec(
